Use a more consistent package structure for Hibernate Search extensions
