--- extensions/hibernate-search-orm-elasticsearch/deployment/src/test/java/io/quarkus/hibernate/search/elasticsearch/test/configuration/CustomMappingFileNotFoundTest.java
@@ -1,31 +1,0 @@
-package io.quarkus.hibernate.search.elasticsearch.test.configuration;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import org.jboss.shrinkwrap.api.ShrinkWrap;
-import org.jboss.shrinkwrap.api.spec.JavaArchive;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-import io.quarkus.runtime.configuration.ConfigurationException;
-import io.quarkus.test.QuarkusUnitTest;
-
-public class CustomMappingFileNotFoundTest {
-
-    @RegisterExtension
-    static final QuarkusUnitTest config = new QuarkusUnitTest().setArchiveProducer(
-            () -> ShrinkWrap.create(JavaArchive.class).addClass(IndexedEntity.class))
-            .withConfigurationResource("application.properties")
-            .overrideConfigKey("quarkus.hibernate-search-orm.elasticsearch.schema-management.mapping-file",
-                    "does-not-exist.json")
-            .assertException(throwable -> assertThat(throwable)
-                    .isInstanceOf(ConfigurationException.class)
-                    .hasMessageContainingAll(
-                            "Unable to find file referenced in 'quarkus.hibernate-search-orm.elasticsearch.schema-management.mapping-file=does-not-exist.json'",
-                            "Remove property or add file to your path"));
-
-    @Test
-    public void test() {
-        // an exception should be thrown
-    }
-}
