--- extensions/hibernate-search-orm-elasticsearch/deployment/src/test/java/io/quarkus/hibernate/search/elasticsearch/test/multiplepersistenceunits/pu3/package-info.java
@@ -1,4 +1,0 @@
-@PersistenceUnit("pu3")
-package io.quarkus.hibernate.search.elasticsearch.test.multiplepersistenceunits.pu3;
-
-import io.quarkus.hibernate.orm.PersistenceUnit;
