--- extensions/hibernate-search-orm-elasticsearch/deployment/src/test/java/io/quarkus/hibernate/search/elasticsearch/test/offline/IndexedEntity.java
@@ -1,17 +1,0 @@
-package io.quarkus.hibernate.search.elasticsearch.test.offline;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
-
-@Entity
-@Indexed
-public class IndexedEntity {
-
-    @Id
-    @GeneratedValue
-    public Long id;
-
-}
