+++ extensions/hibernate-search-orm-elasticsearch/deployment/src/test/java/io/quarkus/hibernate/search/orm/elasticsearch/test/multiplepersistenceunits/defaultpu/package-info.java
@@ -1,0 +1,4 @@
+@PersistenceUnit(PersistenceUnit.DEFAULT)
+package io.quarkus.hibernate.search.orm.elasticsearch.test.multiplepersistenceunits.defaultpu;
+
+import io.quarkus.hibernate.orm.PersistenceUnit;
