+++ extensions/hibernate-search-orm-elasticsearch/deployment/src/test/java/io/quarkus/hibernate/search/orm/elasticsearch/test/multiplepersistenceunits/pu2/package-info.java
@@ -1,0 +1,4 @@
+@PersistenceUnit("pu2")
+package io.quarkus.hibernate.search.orm.elasticsearch.test.multiplepersistenceunits.pu2;
+
+import io.quarkus.hibernate.orm.PersistenceUnit;
